backend/where_i_was/where_i_was/requirements.txt
asgiref==3.3.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
dj-rest-auth==1.1.2
Django==3.1.3
django-braces==1.14.0
django-cors-headers==3.5.0
django-oauth-toolkit==1.3.3
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.1
drf-social-oauth2==1.0.8
idna==2.10
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
urllib3==1.25.11
